+++ data/js/inc/setup/import_key.js
@@ -55,6 +55,7 @@ passbolt.setup.steps = passbolt.setup.steps || {};
         step.elts.$errorFeedback
             .removeClass('hidden')
             .html(errorMessage);
+        passbolt.setup.setActionState('submit', 'enabled');
     };
 
     /* ==================================================================================
@@ -114,9 +115,8 @@ passbolt.setup.steps = passbolt.setup.steps || {};
             .then(function(publicKeyArmored) {
                 passbolt.setup.set('key.publicKeyArmored', publicKeyArmored);
             })
-            .fail(function(error) {
+            .then(null, function(error) {
                 step.onError(error);
-                passbolt.setup.setActionState('submit', 'enabled');
             });
     };
 
@@ -170,7 +170,7 @@ passbolt.setup.steps = passbolt.setup.steps || {};
             .then(function() {
                 def.reject('This key is already used by another user');
             })
-            .fail(function() {
+            .then(null, function() {
                 def.resolve(armoredPrivateKey);
             });
         return def;
@@ -184,11 +184,11 @@ passbolt.setup.steps = passbolt.setup.steps || {};
      */
     step.checkKeyExistRemotely = function (armoredPrivateKey) {
         var def = $.Deferred();
-        return passbolt.request('passbolt.setup.checkKeyExistRemotely', step.data.privateKeyInfo.fingerprint)
+        passbolt.request('passbolt.setup.checkKeyExistRemotely', step.data.privateKeyInfo.fingerprint)
             .then(function() {
                 def.resolve(armoredPrivateKey);
             })
-            .fail(function() {
+            .then(null, function() {
                 def.reject('This key doesn\' match any account.');
             });
         return def;